separate abstract input class to own file
- detect widget dataclasses are now separate from abstract class
- custom_widget simplified and generalised to non-numerical types
